+++ infra/common/schema.py
@@ -80,7 +80,7 @@ class CustomModelArgs(BaseModel):
     resource_name: str
     name: str | None = None
     description: str | None = None
-    base_environment_id: str
+    base_environment_id: str | None = None
     base_environment_version_id: str | None = None
     target_name: str | None = None
     target_type: str | None = None